parsing nearly ok with norminette
