+++ src/syntax_validation.c
@@ -1,13 +1,25 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   syntax_validation.c                                :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: vnafissi <vnafissi@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/05/31 21:27:33 by vnafissi          #+#    #+#             */
+/*   Updated: 2022/05/31 21:28:25 by vnafissi         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minishell.h"
 
-int is_redirection(t_token_type type)
+int	is_redirection(t_token_type type)
 {
-    if (type == RED_INPUT
+	if (type == RED_INPUT
 		|| type == RED_OUTPUT
-        || type == HEREDOC
-        || type == RED_APPEND)
-        return (1);
-    return (0);
+		|| type == HEREDOC
+		|| type == RED_APPEND)
+		return (1);
+	return (0);
 }
 
 int	error_next_is_pipe(t_token *token)
@@ -34,15 +46,15 @@ int	error_next_is_redirection(t_token *token)
 {
 	char	*err;
 
-	err=NULL;
+	err = NULL;
 	if (token->type == RED_INPUT)
-		err="<";
+		err = "<";
 	else if (token->type == RED_OUTPUT)
-		err=">";
+		err = ">";
 	else if (token->type == HEREDOC)
-		err="<<";
+		err = "<<";
 	else if (token->type == RED_APPEND)
-		err=">>";
+		err = ">>";
 	if (err)
 	{
 		ft_putstr_fd("bash: syntax error near unexpected token '", 2);
@@ -55,7 +67,7 @@ int	error_next_is_redirection(t_token *token)
 
 void	check_syntax_errors(t_sh *sh)
 {
-	t_token *token;
+	t_token	*token;
 
 	token = sh->token_lst;
 	while (token)
@@ -78,6 +90,6 @@ void	check_syntax_errors(t_sh *sh)
 		}
 		token = token->next;
 		if (sh->error == 1)
-			return;
+			return ;
 	}
 }
