+++ src/tokens_linked_list.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   tokens_linked_list.c                               :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: vnafissi <vnafissi@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/05/31 21:28:42 by vnafissi          #+#    #+#             */
+/*   Updated: 2022/05/31 21:32:59 by vnafissi         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minishell.h"
 #include "parsing.h"
 
@@ -17,13 +29,12 @@ t_token	*create_token(t_token_type type, char *value)
 
 t_token	*add_back_token(t_token *list, t_token_type type, char *value)
 {
-	t_token *new;
+	t_token	*new;
 	t_token	*tmp;
 
 	new = create_token(type, value);
 	if (!new)
-		return NULL;
-
+		return (NULL);
 	if (!list)
 		return (new);
 	else
@@ -38,11 +49,11 @@ t_token	*add_back_token(t_token *list, t_token_type type, char *value)
 
 t_token	*add_middle_token(t_token *current, t_token_type type, char *value)
 {
-	t_token *new;
+	t_token	*new;
 
 	new = create_token(type, value);
 	if (!new)
-		return NULL;
+		return (NULL);
 	if (!current)
 		return (new);
 	else
@@ -53,10 +64,9 @@ t_token	*add_middle_token(t_token *current, t_token_type type, char *value)
 	return (new);
 }
 
-
-void print_tokens(t_token *li)
+void	print_tokens(t_token *li)
 {
-	while(li != NULL)
+	while (li != NULL)
 	{
 		printf("[%s]", li->value);
 		li = li->next;
@@ -64,78 +74,39 @@ void print_tokens(t_token *li)
 	printf("\n");
 }
 
-
-int count_arg_tokens(t_token *li)
+int	count_arg_tokens(t_token *li)
 {
-	int size = 0;
+	int	size;
 
-	while(li != NULL)
+	size = 0;
+	while (li != NULL)
 	{
 		if (li->type == ARG || li->type == CMD)
 			size++;
 		li = li->next;
 	}
-	return size;
+	return (size);
 }
 
-int list_length(t_token *li)
-{
-	int size = 0;
-
-	while(li != NULL)
-	{
-		size++;
-		li = li->next;
-	}
-	return size;
-}
-
-void	ft_set_null_free_elem(t_token *elem)
-{
-	if (!elem)
-		return ;
-	elem->value = NULL;
-	elem->next = NULL;
-	free(elem);
-}
-
-void	ft_set_null_free_list(t_token **a_list)
-{
-	t_token	*tmp;
-
-	if (!(*a_list))
-		return ;
-	while (*a_list)
-	{
-		tmp = (*a_list)->next;
-		ft_set_null_free_elem(*a_list);
-		*a_list = tmp;
-	}
-}
-
-//void	ft_stack_add_front(t_stack **stack, t_stack *new)
+//void	ft_set_null_free_elem(t_token *elem)
 //{
-//	if (new)
-//	{
-//		new->next = *stack;
-//		*stack = new;
-//	}
+//	if (!elem)
+//		return ;
+//	elem->value = NULL;
+//	elem->next = NULL;
+//	free(elem);
 //}
 
-
-//int	ft_stack_len(t_stack **stack)
+//void	ft_set_null_free_list(t_token **a_list)
 //{
-//	int		i;
-//	t_stack	*tmp;
+//	t_token	*tmp;
 
-//	if (!(*stack))
-//		return (0);
-//	tmp = *stack;
-//	i = 1;
-//	while (tmp->next)
+//	if (!(*a_list))
+//		return ;
+//	while (*a_list)
 //	{
-//		tmp = tmp->next;
-//		i++;
+//		tmp = (*a_list)->next;
+//		ft_set_null_free_elem(*a_list);
+//		*a_list = tmp;
 //	}
-//	return (i);
 //}
